Restructure the tests
